Ran cargo clippy --fix to fix warnings

+++ core/src/util/debug_initializer.rs
@@ -63,12 +63,10 @@ impl InitConfig {
 			let mut config: InitConfig = serde_json::from_str(&config).unwrap();
 			config.path = path;
 
-			if config.reset_on_startup {
-				if data_dir.exists() {
-					warn!("previous 'SD_DATA_DIR' was removed on startup!");
-					fs::remove_dir_all(&data_dir).await.unwrap();
-				}
-			}
+			if config.reset_on_startup && data_dir.exists() {
+   					warn!("previous 'SD_DATA_DIR' was removed on startup!");
+   					fs::remove_dir_all(&data_dir).await.unwrap();
+   				}
 
 			return Some(config);
 		}
